+++ config/api/descriptions.php
@@ -59,5 +59,6 @@ return [
         'GET_category' => 'Return the category summary for a resource',
         'GET_sub_categories' => 'Return the category sub categories summary for a resource',
         'GET_sub_category' => 'Return the category sub category summary for a resource',
+        'GET_years' => 'Return the years summary for a resource',
     ]
 ];